Rename to metaphor.airflow
